backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2
python-magic>=0.4.27

# File analysis
pefile>=2023.2.7
PyPDF2>=3.0.1
yara-python>=4.3.1

# Try to include peepdf for more detailed PDF analysis
# peepdf is not on PyPI, consider manually installing from:
# https://github.com/jesparza/peepdf

# ML
scikit-learn>=1.4.0
joblib>=1.3.2
numpy>=1.26.0
matplotlib>=3.8.2

# Reporting
jinja2>=3.1.3
weasyprint>=60.2

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
python-dotenv>=1.0.1

# Optional: alternative PDF analysis libraries
# pikepdf>=8.11.2
# pdfminer.six>=20221105